Add support for a --no-push option
When making security-related releases, we want to do all release-tagging
related activity locally, without exposing the actual change.  This is
to support a "binary-only" release to give admins the opportunity to
install the upgrade without making details of the vulnerability public.

One an embargo period is over, the local changes may be pushed to
github.

Care must be taken that this does not trigger the usual build process,
as this will result in packages with different checksum values.