Fix a couple of problems with binary releases
When building 2.16, the git information is required.
The old version mistakenly omitted the .gitignore file, so
the source tarball, when unpacked, appeared dirty.  This
triggered the wrong description being embedded in
the jar file's MANIFEST.

This patch fixes this issue, so that all files are included.
Some other minor tweaks are included.